+++ CHANGELOG.md
@@ -8,9 +8,9 @@ Graphics.UI.Threepenny
 
 JavaScript FFI
 
-* Clarify semantics for exceptions.
     * The `UI` monad now supports exception handling by being an instance of the type classes `MonadThrow` and `MonadCatch`.
-    * The function `callFunction` can now throw a `JavaScriptException` exception.
+    * The function `callFunction` can now throw a `JavaScriptException` exception to the Haskell side.
+    * The function `runFunction` now terminates the connection to the browser window whenever the JavaScript code within throws an exception.
 * Clarify semantics of the `disconnect` event. It is now triggered more reliably. [#133][].
 * Remove unnecessary client response when exporting event handlers. [#131][].